Install Express, Rename server.js file to old-server.js
